Added Health System for player, and Basic Shooting AI for enemies
Health system works in a way that initial health is 100 and with each bullet colliding with out player it decreases by "20" and when it reaches "0", "You Lose" UI screen became active.
Shooting AI works similar to shooting by player and they shoot in the direction of player the enemy is facing.

"Exception": Caught one null exception as when player lose it became unactive and gives this error, this is of no worry as after losing game will pause/end condition which will fix this